Add ContactList.java
